# fastapi/api/openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/auth/register/":{"post":{"tags":["auth"],"summary":"Register","operationId":"register_auth_register__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUser"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login/":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_auth_login__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Login"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccessToken"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/store/movies/{movie_uuid}/":{"get":{"tags":["store"],"summary":"Retrieve Movie","operationId":"retrieve_movie_store_movies__movie_uuid___get","parameters":[{"required":true,"schema":{"title":"Movie Id","type":"string","format":"uuid"},"name":"movie_id","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Movie"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/store/movies/":{"get":{"tags":["store"],"summary":"List Movies","operationId":"list_movies_store_movies__get","parameters":[{"required":false,"schema":{"title":"Page","minimum":0.0,"type":"integer","default":0},"name":"page","in":"query"},{"required":false,"schema":{"title":"Page Size","maximum":100.0,"minimum":10.0,"type":"integer","default":10},"name":"page_size","in":"query"},{"required":false,"schema":{"title":"Search","type":"string"},"name":"search","in":"query"},{"required":false,"schema":{"title":"Director","type":"string"},"name":"director","in":"query"},{"required":false,"schema":{"title":"Year","type":"integer"},"name":"year","in":"query"},{"required":false,"schema":{"title":"Genre","type":"string"},"name":"genre","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MoviePaginatedList"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/store/rentals/{rental_uuid}/":{"get":{"tags":["store"],"summary":"Retrieve Rental","operationId":"retrieve_rental_store_rentals__rental_uuid___get","parameters":[{"required":true,"schema":{"title":"Rental Uuid","type":"string","format":"uuid"},"name":"rental_uuid","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Rental"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]},"patch":{"tags":["store"],"summary":"Patch Rental","operationId":"patch_rental_store_rentals__rental_uuid___patch","parameters":[{"required":true,"schema":{"title":"Rental Uuid","type":"string","format":"uuid"},"name":"rental_uuid","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Rental"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/store/rentals/":{"get":{"tags":["store"],"summary":"List Rentals","operationId":"list_rentals_store_rentals__get","parameters":[{"required":false,"schema":{"title":"Page","minimum":0.0,"type":"integer","default":0},"name":"page","in":"query"},{"required":false,"schema":{"title":"Page Size","maximum":100.0,"minimum":10.0,"type":"integer","default":10},"name":"page_size","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RentalPaginatedList"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["store"],"summary":"Create Rental","operationId":"create_rental_store_rentals__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRental"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Rental"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}}},"components":{"schemas":{"AccessToken":{"title":"AccessToken","required":["access_token"],"type":"object","properties":{"access_token":{"title":"Access Token","type":"string"},"token_type":{"title":"Token Type","type":"string","default":"bearer"}}},"CreateRental":{"title":"CreateRental","required":["movie_uuid"],"type":"object","properties":{"movie_uuid":{"title":"Movie Uuid","type":"string","format":"uuid"}}},"CreateUser":{"title":"CreateUser","type":"object","properties":{"username":{"title":"Username","minLength":4,"type":"string"},"email":{"title":"Email","pattern":"^[a-z0-9]+[\\._]?[a-z0-9]+[@]\\w+[.]\\w{2,3}$","type":"string"},"password":{"title":"Password","maxLength":32,"minLength":8,"type":"string"},"password_verification":{"title":"Password Verification","maxLength":32,"minLength":8,"type":"string"}}},"Genre":{"title":"Genre","required":["uuid","name"],"type":"object","properties":{"uuid":{"title":"Uuid","type":"string","format":"uuid"},"name":{"title":"Name","type":"string"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"Login":{"title":"Login","type":"object","properties":{"email":{"title":"Email","pattern":"^[a-z0-9]+[\\._]?[a-z0-9]+[@]\\w+[.]\\w{2,3}$","type":"string"},"password":{"title":"Password","maxLength":32,"minLength":8,"type":"string"}}},"Movie":{"title":"Movie","required":["uuid","title","description","year","director","genres"],"type":"object","properties":{"uuid":{"title":"Uuid","type":"string","format":"uuid"},"title":{"title":"Title","type":"string"},"description":{"title":"Description","type":"string"},"year":{"title":"Year","type":"integer"},"director":{"title":"Director","type":"string"},"genres":{"title":"Genres","type":"array","items":{"$ref":"#/components/schemas/Genre"}}}},"MoviePaginatedList":{"title":"MoviePaginatedList","required":["page","page_size","count","results"],"type":"object","properties":{"page":{"title":"Page","type":"integer"},"page_size":{"title":"Page Size","type":"integer"},"count":{"title":"Count","type":"integer"},"results":{"title":"Results","type":"array","items":{"$ref":"#/components/schemas/Movie"}}}},"Rental":{"title":"Rental","required":["uuid","movie_relationship","rental_date"],"type":"object","properties":{"uuid":{"title":"Uuid","type":"string","format":"uuid"},"movie":{"$ref":"#/components/schemas/Movie"},"rental_date":{"title":"Rental Date","type":"string","format":"date-time"},"return_date":{"title":"Return Date","type":"string","format":"date-time"},"fee":{"title":"Fee","type":"number"}}},"RentalPaginatedList":{"title":"RentalPaginatedList","required":["page","page_size","count","results"],"type":"object","properties":{"page":{"title":"Page","type":"integer"},"page_size":{"title":"Page Size","type":"integer"},"count":{"title":"Count","type":"integer"},"results":{"title":"Results","type":"array","items":{"$ref":"#/components/schemas/Rental"}}}},"User":{"title":"User","required":["uuid"],"type":"object","properties":{"username":{"title":"Username","minLength":4,"type":"string"},"email":{"title":"Email","pattern":"^[a-z0-9]+[\\._]?[a-z0-9]+[@]\\w+[.]\\w{2,3}$","type":"string"},"uuid":{"title":"Uuid","type":"string","format":"uuid"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}